backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.20
alembic==1.14.1
celery==5.4.0
redis==5.2.1
httpx==0.28.1
langchain-core==0.3.58
langchain-openai==0.2.14
langchain-google-genai==2.0.8
langchain-ollama==0.2.3
PyJWT==2.10.1
cryptography==45.0.4
prometheus-client==0.21.1
google-genai>=1.0.0
# Pattern DB (Postgres + pgvector)
psycopg2-binary==2.9.10
pgvector==0.3.6
numpy==2.2.5
